+++ ariadne_future/object_type.py
@@ -1,18 +1,8 @@
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Mapping,
-    Optional,
-    Type,
-    Union,
-    cast,
-)
+from typing import Dict, Optional, Type, Union, cast
 
 from graphql import (
     DefinitionNode,
     GraphQLFieldResolver,
-    GraphQLResolveInfo,
     ObjectTypeDefinitionNode,
     ObjectTypeExtensionNode,
 )
@@ -20,15 +10,14 @@ from graphql import (
 from .base_type import BaseType
 from .dependencies import Dependencies, get_dependencies_from_object_type
 from .types import FieldsDict, RequirementsDict
-from .utils import parse_definition
+from .utils import ResolversMixin, parse_definition
 
 ObjectNodeType = Union[ObjectTypeDefinitionNode, ObjectTypeExtensionNode]
 
 
-class ObjectType(BaseType):
+class ObjectType(BaseType, ResolversMixin):
     __abstract__ = True
     __aliases__: Optional[Dict[str, str]] = None
-    __args__: Optional[Dict[str, Dict[str, str]]]
 
     graphql_name: str
     graphql_type: Union[Type[ObjectTypeDefinitionNode], Type[ObjectTypeExtensionNode]]
@@ -100,18 +89,6 @@ class ObjectType(BaseType):
             )
 
     @classmethod
-    def __validate_aliases__(cls):
-        if not cls.__aliases__:
-            return
-
-        invalid_aliases = set(cls.__aliases__) - set(cls.graphql_fields)
-        if invalid_aliases:
-            raise ValueError(
-                f"{cls.__name__} class was defined with aliases for fields not in "
-                f"GraphQL type: {', '.join(invalid_aliases)}"
-            )
-
-    @classmethod
     def __get_fields__(cls, type_def: ObjectNodeType) -> FieldsDict:
         if not type_def.fields and not (
             isinstance(type_def, ObjectTypeExtensionNode)
@@ -129,71 +106,8 @@ class ObjectType(BaseType):
         return get_dependencies_from_object_type(type_def)
 
     @classmethod
-    def __get_resolvers__(cls):
-        aliases = cls.__aliases__ or {}
-        defined_resolvers = cls.__get_defined_resolvers__()
-
-        used_resolvers = []
-        resolvers = {}
-
-        for field_name in cls.graphql_fields:
-            if aliases and field_name in aliases:
-                resolver_name = aliases[field_name]
-                if resolver_name in defined_resolvers:
-                    used_resolvers.append(resolver_name)
-                    resolvers[field_name] = defined_resolvers[resolver_name]
-                else:
-                    resolvers[field_name] = create_alias_resolver(resolver_name)
-
-            elif field_name in defined_resolvers:
-                used_resolvers.append(field_name)
-                resolvers[field_name] = defined_resolvers[field_name]
-
-        unused_resolvers = [
-            f"resolve_{field_name}"
-            for field_name in set(defined_resolvers) - set(used_resolvers)
-        ]
-        if unused_resolvers:
-            raise ValueError(
-                f"{cls.__name__} class was defined with resolvers for fields not in "
-                f"GraphQL type: {', '.join(unused_resolvers)}"
-            )
-
-        return resolvers
-
-    @classmethod
-    def __get_defined_resolvers__(cls) -> Dict[str, Callable]:
-        resolvers = {}
-        for name in dir(cls):
-            if not name.startswith("resolve_"):
-                continue
-
-            value = getattr(cls, name)
-            if callable(value):
-                resolvers[name[8:]] = value
-
-        return resolvers
-
-    @classmethod
     def __bind_to_schema__(cls, schema):
         graphql_type = schema.type_map.get(cls.graphql_name)
 
         for field_name, field_resolver in cls.resolvers.items():
             graphql_type.fields[field_name].resolve = field_resolver
-
-
-def create_alias_resolver(field_name: str):
-    def default_aliased_field_resolver(
-        source: Any, info: GraphQLResolveInfo, **args: Any
-    ) -> Any:
-        value = (
-            source.get(field_name)
-            if isinstance(source, Mapping)
-            else getattr(source, field_name, None)
-        )
-
-        if callable(value):
-            return value(info, **args)
-        return value
-
-    return default_aliased_field_resolver
